requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
httpx>=0.25.0
aiohttp>=3.9.0
tiktoken>=0.5.1
chromadb>=0.4.18
sqlalchemy>=2.0.23
duckduckgo-search>=4.1.1